Create updateupgrade.py
This script will run update, and upgrade in your debian terminal.
It will begin by giving you a choice of running update, upgrade, or both in sequence. It will also give you the option of exiting.
After either the update or upgrade you will have to push enter to move forward. The results(whether update/upgrade was successful or not) will be logged in the file you have set up as your log file, as well as the date and time for application start and end, and time stamps for every action you perform.